More reliability enhancements to deploy.ps1

+++ deploy.ps1
@@ -1,15 +1,29 @@
 #
-# Powershell v3.0 deployment script for Server 2012 OSes to deploy IIS sites
-# and custom folders directly from TFS build drop folders.
+# Powershell deployment script for Server 2012 OSes to deploy sites and folders directly from
+# TFS build drop folders.
+#
+# To be executed from the server being deployed to. CredSSP authentication must be enabled on the
+# remote machine in order for AD credentials to float over the remoting session. Kerberos
+# authentication (default) will cause access denied errors.
+#
+# James Dunne <james.jdunne@gmail.com>
+# 2013-02-21
+#
+# For usage instructions, see README.md
 #
 
 Param (
-    # Name of PS1 script that creates $Deploy variable, e.g. "Rules\sample.ps1"
+    # Name of PS1 script that creates $Deploy variable
     [Parameter(Mandatory=$true,Position=0)]
-    [string] $ConfigScriptPath
+    [string] $ConfigScriptPath,
+    # Hash of mappings of TFS build name to build numbers
+    $BuildNumbers = @{},
+    # Use -TestMode to disable delete/xcopy
+    [switch] $TestMode
 );
 
-Import-Module WebAdministration;
+# Must stop on all errors:
+$ErrorActionPreference = "Stop";
 
 # Sets appSettings/CurrentEnvironment value
 function Config-Set-CurrentEnvironment ( [System.Xml.XmlDocument]$doc )
@@ -20,7 +34,7 @@ function Config-Set-CurrentEnvironment ( [System.Xml.XmlDocument]$doc )
     foreach ($item in $root.appSettings.add) {
         if ($item.key -eq "CurrentEnvironment") {
             $item.value = $EnvironmentName;
-            "    Updated CurrentEnvironment value = '$EnvironmentName'";
+            Write-Host "    Updated CurrentEnvironment value = '$EnvironmentName'";
         }
     }
 }
@@ -28,11 +42,16 @@ function Config-Set-CurrentEnvironment ( [System.Xml.XmlDocument]$doc )
 # Do various updates to web.config/app.config files:
 function Config-Update ( $path )
 {
+    if ($TestMode) {
+        Write-Host "    TEST MODE: Config-Update path='$path'";
+        return;
+    }
+
     # Find web.config:
     $webConfigPath = [System.IO.Path]::Combine( $path, "web.config" );
     if ( [System.IO.File]::Exists( $webConfigPath ) )
     {
-        "    Updating '$webConfigPath'";
+        Write-Host "    Updating '$webConfigPath'";
 
         [System.Xml.XmlDocument]$doc = new-object System.Xml.XmlDocument;
         $doc.Load($webConfigPath);
@@ -42,20 +61,18 @@ function Config-Update ( $path )
         # Change compilation mode to debug="false"  
         $root."system.web".compilation.debug = "false";
 
-        # Add any other automated web.config customizations here.
-
         # Set CurrentEnvironment:
         Config-Set-CurrentEnvironment $doc;
 
         $doc.Save($webConfigPath);
 
-        "    Updated '$webConfigPath'";
+        Write-Host "    Updated '$webConfigPath'";
     }
 
     # Find *.exe.config files:
     $appConfigFiles = [System.IO.Directory]::GetFiles($path, "*.exe.config");
     foreach ($appConfigPath in $appConfigFiles) {
-        "    Updating '$appConfigPath'";
+        Write-Host "    Updating '$appConfigPath'";
 
         [System.Xml.XmlDocument]$doc = new-object System.Xml.XmlDocument;
         $doc.Load($appConfigPath);
@@ -65,24 +82,32 @@ function Config-Update ( $path )
 
         $doc.Save($appConfigPath);
 
-        "    Updated '$webConfigPath'";
+        Write-Host "    Updated '$webConfigPath'";
     }
 }
 
 # Clears out the $dest folder and xcopies from $src
 function Clear-And-Copy ( $src, $dest )
 {
+    # System folder paths can easily creep in here:
+    if ($dest.StartsWith("C:\", [System.StringComparison]::OrdinalIgnoreCase)) {
+        throw "Will not clear out any folders on C: drive!";
+    }
+
+    if ($TestMode) {
+        Write-Host "    TEST MODE: Clear-And-Copy src='$src' dest='$dest'";
+        return;
+    }
+
     # Clear out the target directory:
-    "    Clearing existing directory: $dest";
+    Write-Host "    Clearing existing directory: $dest";
     if ([System.IO.Directory]::Exists($dest)) {
         # Delete the directory recursively:
         Remove-Item -Path $dest -Recurse -Force;
-
-       #[System.IO.Directory]::Delete($dest, $true);
     }
-    "";
+    Write-Host "";
 
-    "    Copying from '$src' to '$dest'";
+    Write-Host "    Copying from '$src' to '$dest'";
     xcopy $src $dest /E /I /C /Q;
 }
 
@@ -93,8 +118,9 @@ function Deploy-Site ( $ws, $destRelPath, $srcAbsPath )
     if ($destRelPath -ne ".") {
         $siteName = $siteName + "/" + $destRelPath;
     }
-
-    "    Deploying to $siteName";
+    
+    Write-Host "";
+    Write-Host "    Deploying to $siteName";
 
     $dest = [System.IO.Path]::GetFullPath( [System.IO.Path]::Combine($ws.physicalPath, $destRelPath) );
 
@@ -104,13 +130,13 @@ function Deploy-Site ( $ws, $destRelPath, $srcAbsPath )
     # Find the web.config/*.exe.config files and update them accordingly:
     Config-Update $dest;
 
-    "    Deployed to $siteName";
+    Write-Host "    Deployed to $siteName";
 }
 
 # Verifies that a local IIS website exists with the given $name
 function Require-IIS-Website ( $name )
 {
-    $ws = Get-Website -Name $name;
+    $ws = Get-Item "IIS:\Sites\$name";
     if (!$ws) {
         throw "  Could not access IIS data, or site named '$name' does not exist! Try running as Administrator.";
     }
@@ -128,35 +154,49 @@ function Require-Path-Exists ( $path )
     }
 }
 
+function Get-Build-Dir ( $tfsBuildName, $buildRootPath )
+{
+    # Look up the TFS Build Name in the $BuildNumbers param hash:
+    if (!$BuildNumbers.ContainsKey($tfsBuildName)) {
+        Write-Host "  Detecting latest build directory in '$buildRootPath'...";
+
+        $buildDir = (new-object System.IO.DirectoryInfo($buildRootPath)).GetDirectories() |
+            Sort-Object LastWriteTime -descending |
+            Select-Object -first 1;
+
+        if ($buildDir -eq $null) {
+            throw "  Error detecting latest build directory from '$buildRootPath'";
+        }
+
+        Write-Host "  Successfully detected latest build: $buildDir";
+        return $buildDir;
+    } else {
+        $buildDir = $BuildNumbers[$tfsBuildName];
+        Write-Host "  Using specified build: $buildDir";
+        return $buildDir;
+    }
+}
+
 # Process a "WebSites" deployment rule and controls IIS with stop/start commands
 function Deploy-WebSites ( $dp )
 {
     $tfsBuildName = $dp["TFSBuildName"];
 
-    "";
-    "Deploying WebSites from $tfsBuildName...";
+    Write-Host "";
+    Write-Host "Deploying WebSites from $tfsBuildName...";
 
     $buildRootPath = [System.IO.Path]::Combine($BuildsDropPath, $tfsBuildName);
 
-    "  Detecting latest build directory in '$buildRootPath'...";
+    $buildDir = Get-Build-Dir $tfsBuildName $buildRootPath;
 
-    $buildDir = (new-object System.IO.DirectoryInfo($buildRootPath)).GetDirectories() |
-        Sort-Object LastWriteTime -descending |
-        Select-Object -first 1;
-
-    if ($buildDir -eq $null) {
-        throw "  Error detecting latest build directory from '$buildRootPath'";
-    }
-
-    "  Successfully detected latest build: $buildDir";
-    "";
+    Write-Host "";
 
     $buildSitesPath = [System.IO.Path]::Combine($buildRootPath, [System.IO.Path]::Combine($buildDir, "_PublishedWebsites"));
 
     $sites = $dp["WebSites"];
 
     # Check that build folder paths exist (build might not be completed yet):
-    "  Verifying build is complete...";
+    Write-Host "  Verifying build is complete...";
 
     foreach ($site in $sites) {
         # Construct the final source path:
@@ -168,39 +208,91 @@ function Deploy-WebSites ( $dp )
         Require-Path-Exists $site["_SourcePath"];
     }
 
-    "  Verified";
-    "";
-    "  Verifying IIS configuration...";
+    Write-Host "  Verified";
+    Write-Host "";
+    Write-Host "  Verifying IIS configuration...";
+
+    $IISSiteNames = New-Object System.Collections.Generic.HashSet[System.String];
+    $IISAppPoolNames = New-Object System.Collections.Generic.HashSet[System.String];
 
     foreach ($site in $sites) {
         # Require that the target IIS site name exists:
-        $site["_IISSite"] = Require-IIS-Website $site["TargetSite"];
+        $mainSite = Require-IIS-Website $site["TargetSite"];
+        $site["_IISSite"] = $mainSite;
+
+        $tmp = $IISSiteNames.Add($mainSite.Name);
+        $tmp = $IISAppPoolNames.Add($mainSite.applicationPool);
+
+        # NOTE(jsd): Collection includes main site.
+        # TODO(jsd): is it recursive?
+        foreach ($subsite in $mainSite.Collection) {
+            if ($subsite.Name -eq "") { continue; }
+            $tmp = $IISSiteNames.Add($subsite.Name);
+            $tmp = $IISAppPoolNames.Add($subsite.applicationPool);
+        }
     }
     
-    "  Verified";
-    "";
-    "  Stopping IIS sites...";
+    Write-Host "  Verified";
+    Write-Host "";
+    Write-Host "  Stopping IIS sites...";
 
-    foreach ($site in $sites) {
-        $site["_IISSite"].Stop();
+    foreach ($name in $IISSiteNames) {
+        if ($name -eq "") { continue; }
+        Write-Host "    Stopping IIS site '$name'...";
+        try {
+            $iisSite = Get-Website -Name $name;
+            $iisSite.Stop();
+        } catch {
+            Write-Host -ForegroundColor Red "$_";
+        }
+    }
+    foreach ($name in $IISAppPoolNames) {
+        if ($name -eq "") { continue; }
+        Write-Host "    Stopping IIS app pool '$name'...";
+        try {
+            $iisAppPool = Get-Item "IIS:\AppPools\$name";
+            $iisAppPool.Stop();
+        } catch {
+            Write-Host -ForegroundColor Red "$_";
+        }
     }
 
-    "";
-    "  Deploying new code to IIS sites...";
+    Write-Host "  Waiting for IIS sites to stop...";
+    Start-Sleep -Seconds 2;
+
+    Write-Host "";
+    Write-Host "  Deploying new code to IIS sites...";
 
     foreach ($site in $sites) {
         Deploy-Site $site["_IISSite"] $site["TargetFolder"] $site["_SourcePath"];
     }
 
-    "";
-    "  Starting IIS sites...";
+    Write-Host "";
+    Write-Host "  Starting IIS sites...";
     
-    foreach ($site in $sites) {
-        $site["_IISSite"].Start();
+    foreach ($name in $IISAppPoolNames) {
+        if ($name -eq "") { continue; }
+        Write-Host "    Starting IIS app pool '$name'...";
+        try {
+            $iisAppPool = Get-Item "IIS:\AppPools\$name";
+            $iisAppPool.Start();
+        } catch {
+            Write-Host -ForegroundColor Red "$_";
+        }
+    }
+    foreach ($name in $IISSiteNames) {
+        if ($name -eq "") { continue; }
+        Write-Host "    Starting IIS site '$name'...";
+        try {
+            $iisSite = Get-Website -Name $name;
+            $iisSite.Start();
+        } catch {
+            Write-Host -ForegroundColor Red "$_";
+        }
     }
 
-    "";
-    "Deployed WebSites";
+    Write-Host "";
+    Write-Host "Deployed WebSites";
 }
 
 # Handles a "Folders" deployment rule with simple xcopy commands
@@ -208,24 +300,16 @@ function Deploy-Folders ($dp)
 {
     $tfsBuildName = $dp["TFSBuildName"];
 
-    "";
-    "Deploying Folders from $tfsBuildName...";
+    Write-Host "";
+    Write-Host "Deploying Folders from $tfsBuildName...";
 
     $buildRootPath = [System.IO.Path]::Combine($BuildsDropPath, $tfsBuildName);
 
-    "  Detecting latest build directory in '$buildRootPath'...";
-
-    $buildDir = (new-object System.IO.DirectoryInfo($buildRootPath)).GetDirectories() |
-        Sort-Object LastWriteTime -descending |
-        Select-Object -first 1;
-
-    if ($buildDir -eq $null) {
-        throw "  Error detecting latest build directory from '$buildRootPath'";
-    }
-
-    "  Successfully detected latest build: $buildDir";
+    $buildDir = Get-Build-Dir $tfsBuildName $buildRootPath;
 
     $buildTaskPath = [System.IO.Path]::Combine($buildRootPath, $buildDir);
+
+    Write-Host "  Deploying from: $buildTaskPath";
 
     # TODO(jsd): Need a more robust way of detecting when the build completes:
     if ([System.IO.Directory]::GetFiles($buildTaskPath).Count -eq 0) {
@@ -243,8 +327,8 @@ function Deploy-Folders ($dp)
         Config-Update $folder["TargetFolder"];
     }
 
-    "";
-    "Deployed Folders";
+    Write-Host "";
+    Write-Host "Deployed Folders";
 }
 
 
@@ -254,12 +338,14 @@ function Deploy-Folders ($dp)
 #
 
 
+Import-Module WebAdministration;
+
 
 # Execute the deployment configuration script which sets $Deploy:
 $Deploy = $null;
 
 if (!$ConfigScriptPath) {
-    "-ConfigScriptPath parameter is required!";
+    Write-Host "-ConfigScriptPath parameter is required!";
     exit;
 }
 
@@ -267,7 +353,7 @@ if (!$ConfigScriptPath) {
 . $ConfigScriptPath
 
 if ( ($Deploy -eq $null) ) {
-    "$ConfigScriptPath script must set `$Deploy hash!";
+    Write-Host "$ConfigScriptPath script must set `$Deploy hash!";
     exit;
 }
 
@@ -278,13 +364,13 @@ $EnvironmentName = $Deploy["EnvironmentName"];
 $MachineName = Get-Content env:computername;
 
 
-"Deploying for $EnvironmentName environment on $MachineName machine";
+Write-Host "Deploying for $EnvironmentName environment on $MachineName machine";
 
 # Handle all deployment rules in the $Deploy hash:
 foreach ($dp in $Deploy["Deployments"]) {
     if ($dp["Disabled"]) {
-        "";
-        "  Skipping deployment from '$($dp["TFSBuildName"])' because its rule has Disabled = `$true";
+        Write-Host "";
+        Write-Host "  Skipping deployment from '$($dp["TFSBuildName"])' because its rule has Disabled = `$true";
         continue;
     }
 
@@ -294,7 +380,7 @@ foreach ($dp in $Deploy["Deployments"]) {
             Deploy-WebSites $dp;
         }
     } catch {
-        "$_";
+        Write-Host -ForegroundColor Red "$_";
     }
 
     # Deploy Folders:
@@ -303,9 +389,9 @@ foreach ($dp in $Deploy["Deployments"]) {
             Deploy-Folders $dp;
         }
     } catch {
-        "$_";
+        Write-Host -ForegroundColor Red "$_";
     }
 }
 
-"";
-"Deployment complete"
+Write-Host "";
+Write-Host "Deployment complete"
